# api/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "Products_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "summary": "Lista o catálogo. Sem sessão, devolve apenas as peças publicadas.",
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "Products_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{slug}": {
      "get": {
        "operationId": "Products_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "patch": {
        "operationId": "Products_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "summary": "Mudança de `stock` aqui vira ajuste no ledger, assinado por quem pediu —\nver `ProductsService.update`.",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "Products_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "Orders_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "Orders_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "summary": "O checkout da loja cria o pedido sem sessão de admin.",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "patch": {
        "operationId": "Orders_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Trocar a etapa avisa o cliente. O aviso é secundário: se falhar, o status\nmuda do mesmo jeito e a resposta diz por que a mensagem não saiu.",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/custom-requests": {
      "get": {
        "operationId": "CustomRequests_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomRequest"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "CustomRequests"
        ]
      },
      "post": {
        "operationId": "CustomRequests_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRequest"
                }
              }
            }
          }
        },
        "summary": "O formulário da loja envia sem sessão.",
        "tags": [
          "CustomRequests"
        ]
      }
    },
    "/api/v1/custom-requests/{id}": {
      "patch": {
        "operationId": "CustomRequests_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomRequestStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRequest"
                }
              }
            }
          }
        },
        "tags": [
          "CustomRequests"
        ]
      }
    },
...
  },
  "info": {
    "title": "c3dcriativ — API v1",
    "description": "Contrato da API da loja. Preços em CENTAVOS (inteiros) em todos os campos monetários.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductImageEmbed": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "alt": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "alt"
        ]
      },
      "ProductVariantEmbed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "colorHex": {
            "type": "string"
          },
          "priceAdjustment": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "priceAdjustment",
          "stock"
        ]
      },
      "DimensionsEmbed": {
        "type": "object",
        "properties": {
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "depth": {
            "type": "number"
          }
        },
        "required": [
          "width",
          "height",
          "depth"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "category": {
            "enum": [
              "decoracao",
              "colecionaveis",
              "brinquedos",
              "jogos",
              "presentes",
              "personalizados"
            ],
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageEmbed"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantEmbed"
            }
          },
          "material": {
            "type": "string"
          },
          "productionTime": {
            "type": "number"
          },
          "stock": {
            "type": "number",
            "description": "Peças prontas em estoque. Ausente = produção sob demanda."
          },
          "dimensions": {
            "$ref": "#/components/schemas/DimensionsEmbed"
          },
          "weight": {
            "type": "number"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isCustom": {
            "type": "boolean",
            "description": "Peça que aceita personalização (nome, cor, medida) sob encomenda."
          },
          "badge": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "shortDescription",
          "price",
          "category",
          "tags",
          "images",
          "variants",
          "isAvailable",
          "isFeatured",
          "isCustom",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductImageDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1
          },
          "alt": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "url",
          "alt"
        ]
      },
      "ProductVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "colorHex": {
            "type": "string",
            "maxLength": 9
          },
          "priceAdjustment": {
            "type": "number",
            "default": 0
          },
          "stock": {
            "type": "number",
            "default": 0,
            "minimum": 0
          }
        },
        "required": [
          "id",
          "name",
          "priceAdjustment",
          "stock"
        ]
      },
      "DimensionsDto": {
        "type": "object",
        "properties": {
          "width": {
            "type": "number",
            "minimum": 0
          },
          "height": {
            "type": "number",
            "minimum": 0
          },
          "depth": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "width",
          "height",
          "depth"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "minLength": 10
          },
          "shortDescription": {
            "type": "string",
            "minLength": 5
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "originalPrice": {
            "type": "number",
            "minimum": 0
          },
          "category": {
            "type": "string",
            "enum": [
              "decoracao",
              "colecionaveis",
              "brinquedos",
              "jogos",
              "presentes",
              "personalizados"
            ]
          },
          "tags": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "images": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "material": {
            "type": "string"
          },
          "productionTime": {
            "type": "number",
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "minimum": 0
          },
          "dimensions": {
            "$ref": "#/components/schemas/DimensionsDto"
          },
          "weight": {
            "type": "number",
            "minimum": 0
          },
          "isAvailable": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isCustom": {
            "type": "boolean"
          },
          "badge": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "name",
          "description",
          "shortDescription",
          "price",
          "category",
          "tags",
          "images"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "description": {
            "type": "string",
            "minLength": 10
          },
          "shortDescription": {
            "type": "string",
            "minLength": 5
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "originalPrice": {
            "type": "number",
            "minimum": 0
          },
          "category": {
            "type": "string",
            "enum": [
              "decoracao",
              "colecionaveis",
              "brinquedos",
              "jogos",
              "presentes",
              "personalizados"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "images": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "material": {
            "type": "string"
          },
          "productionTime": {
            "type": "number",
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "minimum": 0
          },
          "dimensions": {
            "$ref": "#/components/schemas/DimensionsDto"
          },
          "weight": {
            "type": "number",
            "minimum": 0
          },
          "isAvailable": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isCustom": {
            "type": "boolean"
          },
          "badge": {
            "type": "string",
            "maxLength": 40
          }
        }
      },
...
      "OrderItemEmbed": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "variantId": {
            "type": "string",
            "description": "Id da variação escolhida (`variants[].id`). O carrinho sempre soube qual\nera; o pedido só guardava o NOME dela, que não serve para achar o saldo\ncerto — duas variações podem ter nomes parecidos e o estoque é por id."
          },
          "variantName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "price": {
            "type": "number"
          }
        },
        "required": [
          "productId",
          "name",
          "quantity",
          "price"
        ]
      },
      "CustomerEmbed": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "phone"
        ]
      },
      "AddressEmbed": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "street",
          "number",
          "neighborhood",
          "city",
          "state",
          "zipCode",
          "country"
        ]
      },
      "ExternalRefEmbed": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "shopee",
              "site"
            ]
          },
          "shopId": {
            "type": "string"
          },
          "orderSn": {
            "type": "string"
          },
          "remoteStatus": {
            "type": "string",
            "description": "Último `order_status` visto na origem, para descartar evento atrasado."
          }
        },
        "required": [
          "source",
          "shopId",
          "orderSn"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código curto que a pessoa lê e diz no WhatsApp: C3D-4820."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemEmbed"
            }
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerEmbed"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/AddressEmbed"
          },
          "paymentMethod": {
            "enum": [
              "pix",
              "credit_card",
              "boleto"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "printing",
              "pending",
              "processing",
              "paid",
              "finishing",
              "shipped",
              "delivered",
              "cancelled"
            ],
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "shipping": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "paymentPreferenceId": {
            "type": "string",
            "description": "Preferência do Mercado Pago, quando o pagamento online está ligado."
          },
          "paymentUrl": {
            "type": "string"
          },
          "channel": {
            "enum": [
              "shopee",
              "site"
            ],
            "type": "string"
          },
          "externalRef": {
            "description": "Presente só em pedido importado. O índice único parcial abaixo é o que\ngarante que reimportar o mesmo `order_sn` não crie um segundo pedido.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ExternalRefEmbed"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "items",
          "customer",
          "paymentMethod",
          "status",
          "subtotal",
          "shipping",
          "discount",
          "total",
          "channel",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "pix",
              "credit_card",
              "boleto"
            ]
          },
          "subtotal": {
            "type": "number",
            "minimum": 0
          },
          "shipping": {
            "type": "number",
            "minimum": 0
          },
          "discount": {
            "type": "number",
            "minimum": 0
          },
          "total": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "customer",
          "paymentMethod"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "printing",
              "pending",
              "processing",
              "paid",
              "finishing",
              "shipped",
              "delivered",
              "cancelled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CustomRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "referenceImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "enum": [
              "character",
              "gift",
              "miniature",
              "decoration",
              "functional",
              "other"
            ],
            "type": "string"
          },
          "budget": {
            "type": "number",
            "description": "Em centavos, quando a pessoa informa um teto."
          },
          "deadline": {
            "type": "string"
          },
          "status": {
            "enum": [
              "received",
              "analyzing",
              "quoted",
              "approved",
              "modeling",
              "printing",
              "finished"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "customerName",
          "customerEmail",
          "description",
          "referenceImages",
          "type",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCustomRequestDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "customerEmail": {
            "type": "string",
            "format": "email"
          },
          "customerPhone": {
            "type": "string",
            "maxLength": 20
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000
          },
          "referenceImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "type": "string",
            "enum": [
              "character",
              "gift",
              "miniature",
              "decoration",
              "functional",
              "other"
            ]
          },
          "budget": {
            "type": "number",
            "minimum": 0
          },
          "deadline": {
            "type": "string",
            "maxLength": 60
          }
        },
        "required": [
          "customerName",
          "customerEmail",
          "description"
        ]
      },
      "UpdateCustomRequestStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "received",
              "analyzing",
              "quoted",
              "approved",
              "modeling",
              "printing",
              "finished"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}